+++ src/plugin.cpp
@@ -148,7 +148,7 @@ ENDSTONE_PLUGIN("spark", "0.5.1", SparkPlugin)
             "/spark (tps|cpu|ping|health|healthreport|ht|activity|activitylog|log|tickmonitor|tickmonitoring)<module: "
             "SparkStatusModule> [flags: message]",
             "/spark (profiler|sampler)<module: SparkProfilerModule> "
-            "(start|stop|info|cancel|open|trust-viewer)[action: SparkProfilerAction] [flags: message]")
+            "(start|stop|upload|info|cancel|open|trust-viewer)[action: SparkProfilerAction] [flags: message]")
         .permissions("endstone.command.spark");
 
     permission("endstone.command.spark")